frontend-user-management * added package for adding data into cookie * refactoring for naming * added new function for checking token and get user id
