+++ components/layout/userManagement/SignIn.component.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import Link from 'next/link';
 import { useRouter } from "next/router";
+import { setCookie } from 'cookies-next';
 
 const Login = () => {
   const router = useRouter()
@@ -18,7 +19,7 @@ const Login = () => {
       .then(async data => {
         const responseData = await data.json();
         if (responseData.success) {
-          document.cookie = `token=${responseData.token}`;
+          setCookie('token', responseData.token);
 
           router.push('/');
         }
