+++ helper/generateAccessToken.ts
