+++ package.json
@@ -7,6 +7,7 @@
   },
   "dependencies": {
     "bcrypt": "^5.1.0",
+    "cookies-next": "^2.1.1",
     "crypto": "^1.0.1",
     "gridfs-stream": "^1.1.1",
     "jsonwebtoken": "^8.5.1",
